+++ R-Type/server/includes/Game/Manager.hpp
@@ -50,7 +50,7 @@ namespace NGame
             void onNotifyUsersComponentRemoved(const std::vector<std::shared_ptr<NGame::User>> &, uint64_t);
             void onNotifyUsersComponentAdded(const std::vector<std::shared_ptr<NGame::User>> &, const std::shared_ptr<NGame::Component>&);
     		void onNotifyUserGainScore(const Peer &, uint64_t, const std::string &, uint64_t);
-    		void onNotifyTimeElapsedPing(const Peer &, double);
+            void onNotifyTimeElapsedPing(const Peer &, float);
 
             void setListener(NGame::Manager::OnManagerEvent *listener);
     		void terminatedGame(const std::shared_ptr<NGame::Game>& game);
